+++ client/nodejs/client.js
@@ -9,7 +9,7 @@ module.exports = Client = function(options) {
     process.EventEmitter.call(this);
 
     this.options = utils.merge({
-        host : '127.0.0.1',
+        host : 'localhost',
         port : 4000,
         ssl  : false,
 
@@ -26,13 +26,10 @@ module.exports = Client = function(options) {
     this._io = new IO(this.options);
 
     this._io.on('message', this.router.dispatch.bind(this.router));
-    this._io.on('error', this._onError.bind(this));
-
-    process.on('exit', function() {
-        process.nextTick(function(){
-            this.disconnect();
-        }.bind(this));
-    }.bind(this));
+	this._io.on('error', this._onError.bind(this));
+	this._io.on('disconnect', (function() {
+		this.emit('disconnect');
+	}).bind(this));
 };
 
 Client._statuses = {
@@ -172,3 +169,10 @@ Client.prototype.flushMessageQueue = function() {
 	this._io.send(this._messageQueue);
     this._messageQueue = [];
 };
+/*          { "subscribers" : 128 },
+	            { "subscribers" : 256 },
+	            { "subscribers" : 512 },
+	            { "subscribers" : 1024 },
+	            { "subscribers" : 2048 },
+	            { "subscribers" : 4096 },
+	            { "subscribers" : 8192 }*/